requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
pandas==2.1.3
pytz==2023.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
alembic==1.12.1
pydantic==2.4.2
pydantic-settings==2.0.3
python-dateutil==2.8.2 